# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Bowtie API
  version: 0.1.0
  description: Live TV streaming API for HDHomeRun tuners.
paths:
...
  /api/v1/iptv/{key}/stream/{channelId}:
    parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      - name: channelId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: quality
        in: query
        schema:
          type: string
          enum: [original, high, medium, low]
    get:
      summary: Start or join a channel as the feed account; redirects to its HLS playlist
      operationId: iptvStream
      tags: [iptv]
      responses:
        '302':
          description: Redirect to /api/v1/stream/{viewerId}/index.m3u8?token=…
        '403':
          description: Blocked by parental controls
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          description: Account limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserLimitError'
        '503':
          description: All tuners in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TunersBusyError'
...
  /api/v1/channels:
    get:
      summary: List enabled channels for the viewer
      operationId: listChannels
      tags: [channels]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Enabled channels with optional EPG logo
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ViewerChannel'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/stream/{viewerId}/{segment}:
    parameters:
      - name: viewerId
        in: path
        required: true
        schema:
          type: string
      - name: segment
        in: path
        required: true
        description: |
          A session file: rendition playlist (v<height>.m3u8, v<height>_vtt.m3u8,
          aac<i>.m3u8, ac3<i>.m3u8 — fetching one heartbeats the viewer), MPEG-TS
          segment (<rendition>_#####.ts) or WebVTT segment (v<height><n>.vtt,
          served with an X-TIMESTAMP-MAP header). Other names return 400.
        schema:
          type: string
          pattern: '^((v\d{3,4}|v\d{3,4}_vtt|aac[0-2]|ac3[0-2])\.m3u8|(v\d{3,4}|aac[0-2]|ac3[0-2])_\d{5}\.ts|v\d{3,4}\d+\.vtt)$'
      - name: token
        in: query
        required: true
        schema:
          type: string
    get:
      summary: HLS rendition playlist or segment
      operationId: getSegment
      tags: [stream]
      responses:
        '200':
          description: Rendition playlist, MPEG-TS segment or WebVTT segment
          content:
            application/vnd.apple.mpegurl:
              schema:
                type: string
            video/mp2t:
              schema:
                type: string
                format: binary
            text/vtt:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    UserLimitError:
      type: object
      required: [error, code, kind, limit]
      properties:
        error:
          type: string
          example: Your account can use 1 tuner at a time. Stop another channel first.
        code:
          type: string
          enum: [user_limit]
        kind:
          type: string
          enum: [streams, tuners]
        limit:
          type: integer
...
    ViewerChannel:
      type: object
      required: [id, guideNumber, name, logoUrl, reception, favorite]
      properties:
        favorite:
          type: boolean
          description: The caller has starred this channel
        id:
          type: integer
          format: int64
        guideNumber:
          type: string
        name:
          type: string
        logoUrl:
          type: string
          description: Icon URL from mapped EPG channel, or empty
        reception:
          type: string
          enum: [ok, noSignal, unknown]
          description: |
            The channel's last tune outcome, learned from real tunes (the
            HDHomeRun reports signal only for a channel it is tuned to).
            "noSignal" means the tuner last answered 806/807; "unknown" means
            not tuned since the server started.
        receptionCheckedAt:
          type: string
          format: date-time
          description: When reception was last learned; absent when unknown.
...
    ViewerInfo:
      type: object
      required: [id, username, lastSeen]
      properties:
        id:
          type: string
        username:
          type: string
        lastSeen:
          type: string
          format: date-time
    SessionInfo:
      type: object
      required: [id, channelId, channelName, key, videoCodec, profile, backend, viewers, startedAt]
      properties:
        id:
          type: string
        channelId:
          type: integer
          format: int64
        channelName:
          type: string
        key:
          type: string
        videoCodec:
          type: string
        profile:
          type: string
        backend:
          type: string
        viewers:
          type: array
          items:
            $ref: '#/components/schemas/ViewerInfo'
        startedAt:
          type: string
          format: date-time
    TunersBusyError:
      type: object
      required: [error, sessions]
      description: |
        Returned when no tuner is free. For non-admin callers, `sessions`
        includes only sessions on enabled channels; admins see all sessions.
      properties:
        error:
          type: string
          example: all tuners in use
        sessions:
          type: array
          items:
            $ref: '#/components/schemas/SessionInfo'
        otherInUse:
          type: integer
          description: |
            Tuners busy for something other than Bowtie (for example Plex):
            tuners the HDHomeRun reports streaming minus the ones Bowtie's own
            streams hold.
  responses:
    BadRequest:
      description: Invalid request body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication failed or missing
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated but not permitted (e.g. non-admin)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'